+++ src/PSRule.Types/PipelineException.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 using System.Runtime.Serialization;
+using PSRule.Runtime;
 
 namespace PSRule;
 
@@ -25,12 +26,35 @@ public abstract class PipelineException : Exception
     /// <summary>
     /// Initialize a new instance of a PSRule exception.
     /// </summary>
+    protected PipelineException(EventId eventId, string message)
+        : base(message)
+    {
+        EventId = eventId;
+    }
+
+    /// <summary>
+    /// Initialize a new instance of a PSRule exception.
+    /// </summary>
     protected PipelineException(string message, Exception innerException)
         : base(message, innerException) { }
 
     /// <summary>
     /// Initialize a new instance of a PSRule exception.
     /// </summary>
+    protected PipelineException(EventId eventId, string message, Exception innerException)
+        : base(message, innerException)
+    {
+        EventId = eventId;
+    }
+
+    /// <summary>
+    /// Initialize a new instance of a PSRule exception.
+    /// </summary>
     protected PipelineException(SerializationInfo info, StreamingContext context)
         : base(info, context) { }
+
+    /// <summary>
+    /// The event identifier for the exception.
+    /// </summary>
+    public EventId? EventId { get; }
 }